[#31] refactor: 박스에서 사용하는 drop API 훅으로 분리
